Inject middleware collection into HttpHandlerStackFactory::create()
