+++ src/Services/ArrayCollection.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace App\Services;
+
+class ArrayCollection
+{
+    /**
+     * @var array
+     */
+    private $collection;
+
+    public function __construct(array $collection = [])
+    {
+        $this->collection = $collection;
+    }
+
+    public function get(): array
+    {
+        return $this->collection;
+    }
+}
